pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genomics-assets"
version = "0.1.0"
description = "Shared reference and QC asset management for the IZKF genomics facility."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6",
]

[project.scripts]
genomics-assets = "genomics_assets.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

